requirements.txt
streamlit
boto3
pdfplumber
PyPDF2
python-docx
python-pptx
requests
python-dotenv
